refactor(dir): ♻ convert all surrogate keys to longs

+++ services/Directory/FilterLists.Directory.Application/Commands/CreateList.cs
@@ -12,7 +12,7 @@ public static class CreateList
     public record Command(string Name,
         IEnumerable<FilterListViewUrl> ViewUrls,
         string? Description = default,
-        int? LicenseId = default,
+        long? LicenseId = default,
         Uri? HomeUrl = default,
         Uri? OnionUrl = default,
         Uri? PolicyUrl = default,
