+++ services/Directory/FilterLists.Directory.Infrastructure/Persistence/Commands/EntityTypeConfigurations/FilterListViewUrlTypeConfiguration.cs
@@ -14,6 +14,6 @@ internal class FilterListViewUrlTypeConfiguration : IEntityTypeConfiguration<Fil
         var nr = new SnakeCaseNameRewriter(CultureInfo.InvariantCulture);
 
         builder.ToTable($"{nr.RewriteName(nameof(Queries.Entities.FilterListViewUrl))}s");
-        builder.Property<int>(nameof(Queries.Entities.FilterListViewUrl.Id));
+        builder.Property<long>(nameof(Queries.Entities.FilterListViewUrl.Id));
     }
 }
